add name to processes function, when available
