+++ src/libnetdata/sanitizers/chart_id_and_name.h
@@ -1,0 +1,14 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#ifndef NETDATA_CHART_ID_AND_NAME_H
+#define NETDATA_CHART_ID_AND_NAME_H
+
+#include "../libnetdata.h"
+
+void netdata_fix_chart_id(char *s);
+void netdata_fix_chart_name(char *s);
+
+void sanitize_chart_meta(char *buf);
+STRING *sanitize_chart_meta_string(STRING *s);
+
+#endif //NETDATA_CHART_ID_AND_NAME_H
